Was using garbage memory.

+++ src/unixtls.cpp
@@ -255,7 +255,7 @@ extern "C" int pthread_join (pthread_t thread,
       reinterpret_cast<pthread_join_function>
       (reinterpret_cast<intptr_t>(dlsym (RTLD_NEXT, fname)));
 
-   fprintf(stderr, "> pthread_join called (%d)\n", (uint32_t) thread);
+   fprintf(stderr, "> pthread_join called (%lu)\n", thread);
 
    return -1;
